refactor: converter 폴더 이동 및 서비스 어노테이션 제거
