+++ src/main/java/com/gamemoonchul/application/converter/PostConverter.java
@@ -8,8 +8,6 @@ import org.springframework.stereotype.Service;
 
 @Service
 public class PostConverter {
-
-
     public static Post requestToEntity(PostUploadRequest request, Member member) {
         Post entity = Post.builder()
                 .title(request.title())
